Click_and_Position/qidian/requirements.txt
Pillow
opencv-python
pyyaml
numpy